+++ documentation/breezemap.html
@@ -201,7 +201,7 @@
             <div class="teidiv1" id="options">
                <h3><span class="headingNumber">4.4. </span><span class="head">Options you can pass to the map in JavaScript</span></h3>
                <p>Here's an example of initializing the map in JS:</p>
-               <pre id="index.xml-eg-d33e302" class="pre_eg">holLayer = new hol.VectorLayer(map, featureJSONFile, {<br/> allowDrawing: true,<br/> allowUserTracking: true,<br/> allFeaturesTaxonomy: true,<br/> timelinePanZoom: false,<br/> initialTaxonomyId: 'holTaxonomyAll'<br/> });</pre>
+               <pre id="index.xml-eg-d33e302" class="pre_eg">holLayer = new VectorLayer(map, featureJSONFile, {<br/> allowDrawing: true,<br/> allowUserTracking: true,<br/> allFeaturesTaxonomy: true,<br/> timelinePanZoom: false,<br/> initialTaxonomyId: 'holTaxonomyAll'<br/> });</pre>
                <p>Among the initialization parameters you can set are:</p>
                <ul>
                   <li class="item"><span class="ident">allowDrawing</span>: A boolean specifying whether drawing tools should be provided to allow users to
