+++ templates/BreezeMap.html
@@ -1,84 +1,90 @@
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
+<head>
     <meta content="width=device-width, initial-scale=1.0" name="viewport"/>
-      <meta charset="UTF-8"/>
-      <title>[Map Title]</title>
-        <link rel="stylesheet" href="ol/ol.css" type="text/css"/>
-        <link rel="stylesheet" href="css/hcmc_ol.css" type="text/css"/>
-      
-        <script src="ol/ol.js" type="text/ecmascript"></script>
-<!--   NOTE: WWe're using an unminimized form of the BreezeMap Javascript 
-       for testing purposes. Change this for a mature live project. -->
-        <script src="js/hcmc_ol.js" type="text/ecmascript"></script>
-      
-    </head>
-    <body>
-      
-        <div id="map" class="map"></div>
+    <meta charset="UTF-8"/>
+    <title>[Map Title]</title>
+    <link rel="stylesheet" href="ol/ol.css" type="text/css"/>
+    <link rel="stylesheet" href="css/hcmc_ol.css" type="text/css"/>
 
-       <div id="splash">
-         <p>Created with <a href="https://github.com/martindholmes/BreezeMap">BreezeMap</a>.</p>
-       </div>
-      
-      <script type="text/ecmascript">
-       
-      /* Initial setup of the map with default controls and
-         Open Street Maps tiles. Change these settings as you
-         need to. */
-         
-      //Globals...
-      var mapCentre = [-123.312214, 48.463199];
-      var view = null, myLayers = [], layerOSM = null, featureJSONFile = 'js/[featureJSONFile].json', holLayer = null;
-         
-      function initBreezeMap(){
-      
+    <script src="ol/ol.js" type="text/ecmascript"></script>
+    <!--   NOTE: WWe're using an unminimized form of the BreezeMap Javascript
+           for testing purposes. Change this for a mature live project. -->
+    <script src="js/constants.js" type="module"></script>
+    <script src="js/util.js" type="module"></script>
+    <script src="js/vectorlayer.js" type="module"></script>
+
+</head>
+<body>
+
+<div id="map" class="map"></div>
+
+<div id="splash">
+    <p>Created with <a href="https://github.com/martindholmes/BreezeMap">BreezeMap</a>.</p>
+</div>
+
+<script type="text/ecmascript">
+
+    /* Initial setup of the map with default controls and
+       Open Street Maps tiles. Change these settings as you
+       need to. */
+
+    //Globals...
+    var mapCentre = [-123.312214, 48.463199];
+    var view = null, myLayers = [], layerOSM = null, featureJSONFile = 'js/[featureJSONFile].json', holLayer = null;
+
+    function initBreezeMap() {
+
         //Create a view centred on our area.
         view = new ol.View({
-          projection: 'EPSG:3857',
-          center: ol.proj.fromLonLat(mapCentre),
-          maxZoom: 22,
-          numZoomLevels: 21,
-          zoom: 17
+            projection: 'EPSG:3857',
+            center: ol.proj.fromLonLat(mapCentre),
+            maxZoom: 22,
+            numZoomLevels: 21,
+            zoom: 17
         });
-        
+
         //Create a tile layer using our own built OSM tileset.
         layerOSM = new ol.layer.Tile({
-          source: new ol.source.OSM({
-            crossOrigin: 'anonymous',
-            numZoomLevels: 21,
-            maxZoom: 22
-          })
+            source: new ol.source.OSM({
+                crossOrigin: 'anonymous',
+                numZoomLevels: 21,
+                maxZoom: 22
+            })
         });
-        
+
         myLayers.push(layerOSM);
-        
+
         //Create a map using our new layer.
         map = new ol.Map({
-          layers: myLayers,
-          renderer: 'canvas',
-          target: 'map',
-          interactions: ol.interaction.defaults().extend([
-             	    new ol.interaction.DragRotateAndZoom()
-             	  ]),
-          view: view
+            layers: myLayers,
+            renderer: 'canvas',
+            target: 'map',
+            interactions: ol.interaction.defaults().extend([
+                new ol.interaction.DragRotateAndZoom()
+            ]),
+            view: view
         });
-      
-  
+
+
 //Now let's create a hol vector layer for all our features.
 //You may wish to turn off some of these features for a live map.
-        holLayer = new hol.VectorLayer(map, featureJSONFile, {allowTaxonomyEditing: true, 
-                                                              allowDrawing: true,
-                                                              allowUpload: true,
-                                                              allowUserTracking: true,
-                                                              allFeaturesTaxonomy: true});
-      }
-      
-      window.addEventListener('load', function (event){initBreezeMap();});
-      
-    
-    </script>
-      
-       
-    </body>
+        holLayer = new VectorLayer(map, featureJSONFile, {
+            allowTaxonomyEditing: true,
+            allowDrawing: true,
+            allowUpload: true,
+            allowUserTracking: true,
+            allFeaturesTaxonomy: true
+        });
+    }
+
+    window.addEventListener('load', function (event) {
+        initBreezeMap();
+    });
+
+
+</script>
+
+
+</body>
 </html>
