+++ testdata/campus.html
@@ -1,115 +1,130 @@
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
+<head>
     <meta content="width=device-width, initial-scale=1.0" name="viewport"/>
-      <meta charset="UTF-8"/>
-      <title>UVic Campus</title>
-        <link rel="stylesheet" href="ol/ol.css" type="text/css"/>
-        <link rel="stylesheet" href="css/hcmc_ol.css" type="text/css"/>
-      
-<!--   NOTE: We're using the debug version of OL for testing purposes.
-          Switch to the compiled version for a live site.   -->
-        <script src="ol/ol.js" type="text/ecmascript"></script>
-        <script src="js/hcmc_ol.js" type="text/ecmascript"></script>
-      <style type="text/css">
-<!--        
-        #opacitySlider{
-          position: absolute;
-          top: 5em;
-          left: 3em;
-          z-index: 100;
-          display: block;
-          background-color: rgba(0, 60, 136, 0.5);
-          color: #ffffff;
+    <meta charset="UTF-8"/>
+    <title>UVic Campus</title>
+    <link rel="stylesheet" href="ol/ol.css" type="text/css"/>
+    <link rel="stylesheet" href="css/hcmc_ol.css" type="text/css"/>
+
+    <!--   NOTE: We're using the debug version of OL for testing purposes.
+              Switch to the compiled version for a live site.   -->
+    <script src="ol/ol.js" type="text/ecmascript"></script>
+    <script src="js/constants.js" type="module"></script>
+    <script src="js/util.js" type="module"></script>
+    <script src="js/vectorlayer.js" type="module"></script>
+    <style type="text/css">
+        <!--
+        #opacitySlider {
+            position: absolute;
+            top: 5em;
+            left: 3em;
+            z-index: 100;
+            display: block;
+            background-color: rgba(0, 60, 136, 0.5);
+            color: #ffffff;
         }
-        #opacitySlider:hover{
-          background-color: rgba(0, 60, 136, 0.7);
+
+        #opacitySlider:hover {
+            background-color: rgba(0, 60, 136, 0.7);
         }
+
         .ol-rotate {
-          left: 0.5em;
-          top: 6em;
-          transition: opacity 0.25s linear 0s, visibility 0s linear 0s;
-          width: 1.375em;
+            left: 0.5em;
+            top: 6em;
+            transition: opacity 0.25s linear 0s, visibility 0s linear 0s;
+            width: 1.375em;
         }
+
         -->
-      </style>
-    </head>
-    <body>
-      
-        <div id="map" class="map"></div>
-      <form>
-      <input class="ol-control ol-unselectable" id="opacitySlider" title="Change the base map layer opacity" type="range" step="0.01" max="1" min="0" value="0.6" />
-      </form>
-       
-       <div id="splash">
-         <p>Welcome to the HCMC's UVic Campus Map. This project is a pilot/test environment 
-           for the development of the HCMC's <a href="https://github.com/martindholmes/BreezeMap">BreezeMap</a> project. </p>
-       </div>
-      
-      <script type="text/ecmascript">
-      <!-- 
-      /* Initial setup of the map with default controls using Tileset created by HCMC from 
-         Open Street Maps dataset. */
-         
-      //Globals...
-      var mapCentre = [-123.312214, 48.463199];
-      var map = null, view = null, myLayers = [], layerXYZ = null, featureJSONFile = 'js/campus.json', holLayer = null;
-         
-      function initCampusMap(){
-      
+    </style>
+</head>
+<body>
+
+<div id="map" class="map"></div>
+<form>
+    <input class="ol-control ol-unselectable" id="opacitySlider" title="Change the base map layer opacity" type="range"
+           step="0.01" max="1" min="0" value="0.6"/>
+</form>
+
+<div id="splash">
+    <p>Welcome to the HCMC's UVic Campus Map. This project is a pilot/test environment
+        for the development of the HCMC's <a href="https://github.com/martindholmes/BreezeMap">BreezeMap</a> project.
+    </p>
+</div>
+
+<script type="module">
+
+    /* Initial setup of the map with default controls using Tileset created by HCMC from
+       Open Street Maps dataset. */
+
+    //Globals...
+    import {VectorLayer} from "../js/vectorlayer.js";
+
+    var mapCentre = [-123.312214, 48.463199];
+    var map = null, view = null, myLayers = [], layerXYZ = null, featureJSONFile = 'js/campus.json', holLayer = null;
+
+    function initCampusMap() {
+
         //Create a view centred on our area.
         view = new ol.View({
-          projection: 'EPSG:3857',
-          /*center:  [-13726849,6184454],*/
-          center: ol.proj.fromLonLat(mapCentre),
-          maxZoom: 22,
-          numZoomLevels: 21,
-          zoom: 17
+            projection: 'EPSG:3857',
+            /*center:  [-13726849,6184454],*/
+            center: ol.proj.fromLonLat(mapCentre),
+            maxZoom: 22,
+            numZoomLevels: 21,
+            zoom: 17
         });
-        
+
         //Create a tile layer using our own built OSM tileset.
         layerXYZ = new ol.layer.Tile({
-          source: new ol.source.XYZ({
-            url: 'https://hcmc.uvic.ca/people/greg/mapnik/osm/tiles/{z}/{x}/{y}.png',
-            /*crossOrigin: 'anonymous',*/
-            numZoomLevels: 21,
-            maxZoom: 22
-          })/*,
+            source: new ol.source.XYZ({
+                url: 'https://hcmc.uvic.ca/people/greg/mapnik/osm/tiles/{z}/{x}/{y}.png',
+                /*crossOrigin: 'anonymous',*/
+                numZoomLevels: 21,
+                maxZoom: 22
+            })/*,
           extent: [corner1[0], corner1[1], corner2[0], corner2[1]]*/
         });
-        
+
         myLayers.push(layerXYZ);
-        
+
         //Create a map using our new layer.
         map = new ol.Map({
-          layers: myLayers,
-          renderer: 'canvas',
-          target: 'map',
-          interactions: ol.interaction.defaults.defaults().extend([
-             	    new ol.interaction.DragRotateAndZoom()
-             	  ]),
-          view: view
+            layers: myLayers,
+            renderer: 'canvas',
+            target: 'map',
+            interactions: ol.interaction.defaults.defaults().extend([
+                new ol.interaction.DragRotateAndZoom()
+            ]),
+            view: view
         });
-      
+
 //Set up the opacity control
-        document.getElementById('opacitySlider').onchange = function(){layerXYZ.setOpacity(parseFloat(this.value));};
+        document.getElementById('opacitySlider').onchange = function () {
+            layerXYZ.setOpacity(parseFloat(this.value));
+        };
         document.getElementById('opacitySlider').onchange();
-  
+
 //Now let's create a hol vector layer for all our features.
-        holLayer = new hol.VectorLayer(map, featureJSONFile, {allowTaxonomyEditing: true, 
-                                                              allowDrawing: true,
-                                                              allowUpload: true,
-                                                              allowUserTracking: true,
-                                                              allFeaturesTaxonomy: true,
-                                                              timelinePanZoom: true,
-                                                              msPlayInterval: 3000});
-      }
-      
-      window.addEventListener('load', function (event){initCampusMap();});
-      
-    -->
-    </script>
-      
-       
-    </body>
+        holLayer = new VectorLayer(map, featureJSONFile, {
+            allowTaxonomyEditing: true,
+            allowDrawing: true,
+            allowUpload: true,
+            allowUserTracking: true,
+            allFeaturesTaxonomy: true,
+            timelinePanZoom: true,
+            msPlayInterval: 3000
+        });
+    }
+
+    window.addEventListener('load', function (event) {
+        initCampusMap();
+    });
+
+
+</script>
+
+
+</body>
 </html>
